Declare `apollo-server-core` in `devDependencies`.
This project `apollo-server-plugin-base` uses interfaces that are
defined in `apollo-server-core`, however we don't declare a dependency on
`apollo-server-core` right now.

We can fix this, by moving the types to places where we won't create cycles
will help - we intentionally avoided directly depending on
`apollo-server-core` from the `apollo-server-plugin-base` to allow
implementers of plugins from needing to directly depend on a specific
version of Apollo Server.

We should probably create a peer dependency within
`apollo-server-plugin-base` which depends on a range of versions though, but
that's YTBD.

For now this makes this work within this repository and should let tests
pass.