Update LICENSE.md for new `fast-json-stable-stringify`.
As of https://github.com/apollographql/apollo-server/pull/2065, we no longer
use `json-stable-stringify`, which pulled in `jsonify` — a dependency which
is licensed under a public-domain license which is not agreeable with some
organizations.  The `fast-json-stable-stringify` license does not rely on
`jsonify`.

This swaps the package being licensed, but the license itself is the same
(MIT).