Stop using Lerna `--exact` flag for publishing. (#2)
I'm not sure if the `--exact` flag was ever what we wanted.  Specifically,
the `--exact` flag publishes every package without semantic version symbols
in all of its `package.json` dependencies (e.g. prod, dev, optional).

This means that if any package in this repository depends on another
package, all related packages must be republished when a new version of one
package comes out, even if there are no other changes to the others.

This means we're essentially over-publishing.  My hunch is that, at one
point, when we were trying to publish early-access/preview releases of some
packages, we used the `--exact` flag in combination with general
`prerelease` and `--canary` flags and this has remained as a historical
artifact.