Partially revert f74e10eea31a9093168811b8264cdf12bbd72421.
This partially reverts the changes in f74e10eea31a9093168811b8264cdf12bbd72421 which
changed from using the native `Object.entries` and `Object.values` methods
provided in ES2017 to use `objectEntries` and `objectValues` helpers which
mock that functionality for ES2016.

As we've now dropped support for Node.js and bumped the compilation target
in #4, those abstractions shouldn't be necessary anymore.

This preserves the rest of the changes in the aforementioned commit, such as
the file moves from `federation` to `composition`.