Fix unresolved dependency for `apollo-graphql`.
Unfortunately, the package is provided outside of this repo during
development, and `npm install` takes its hints from the `node_modules`
directory when updating its `package-lock.json`.  Since the package it
originally prescribed to be installed in `node_modules` is hoisted out by
the devkit, it can't resolve it when it takes a fresh inventory of the
`node_modules` directory when updating the `package-lock.json`.

Relates to #31

cc @martijnwalraven