Support for treating root type definintions as extensions (#6)
This commit introduces a way to handle root operation types. 
It's a lot to explain in a commit message, but the comments and 
tests should illustrate the expected behavior well.

In short, we want to transform the schema before composing in
order for the behavior to match our expectations when root types
are composed. This is because root type names can either be specified by
the user or defaulted by graphql.