Publish federation package to GCS on release (#45)
This commit provides a temporary solution to the problem
of making `apollo-federation` accessible internally.

For now, the solution is to push the tarball to GCS
whenever we release.

In order to successfully push to GCS, the user must
specify a path to the GCS json key via the
FEDERATION_PUBLISHER_GCS_KEYFILE env variable. This key
can be obtained through the GCS console and belongs to the
apollo-platform-repo-publisher service account.