Use `pretty-format` `Plugin` type instead of outdated `SnapshotSerializerPlugin`
Snapshot serializers in Jest are actually `pretty-format` plugins: https://github.com/facebook/jest/tree/master/packages/pretty-format#usage-in-jest

`Expect.addSnapshotSerializer` and the `SnapshotSerializerPlugin` type in `@types/jest` only supports old-style plugins however, not the new API introduced with Jest 21. Because some of our serializers rely on that new API, we need to override `Expect.addSnapshotSerializer` and use the `pretty-format` types directly to allow that.

This should ultimately be fixed either in `@types/jest` or in Jest itself,  which is in a transition to convert the codebase to TypeScript.