Use a special environment variable for testing rather than `NODE_ENV`.
When I introduced 4313dee912fe1ddbe0b1a2cb46192c86f23c4dea in #259, it was
in an effort to simplify the testing story for mocking a memoized global
function within the operation registry's `common.ts` functionality.

While I think that works still makes sense, my use of a check of whether
the `process.env.NODE_ENV` variable was set to `test` or not (a mode that
Jest, and many other test runners, sets automatically) is problematic for
anyone who actually tests their own application's connectivity with our
cloud service since it changes the URL that we poll to a test-only host that
allows us to test network functionality by mocking specific URLs behavior
(that tool is called `nock`!).

This mostly keeps the behavior that the above commit introduced, but changes
it to leverage a special environment variable that is only true within our
infrastructure's tests by virtue of a `setupFiles` property in the
`jest.config.js`.

In order to do that, I had to augment the way that `setupFiles` was set
since we only need this environment variable on this one package but current
inherit from the root `jest.config.base.js` of this monorepo.  I also chose
that approach since I suspect, eventually, this code will no longer live in
this monorepo, but will become a part of a different home, though will still
necessitate that override to pass its test suite!